requirements.txt
torch
torchvision
Pillow
ipython
huggingface_hub
openslide-python
opencv-python
matplotlib
numpy
pandas
tqdm
torchmetrics
torchmil
timm
scipy
scikit-learn
h5py